# generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ReviewCreateRequest": {
				"properties": {
					"listingId": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"minimum": 1,
						"maximum": 5
					},
					"comment": {
						"type": "string"
					}
				},
				"required": [
					"listingId",
					"rating"
				],
				"type": "object",
				"additionalProperties": true
			},
			"PaginatedResponse_any_": {
				"properties": {
					"data": {
						"items": {},
						"type": "array"
					},
					"meta": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"limit",
							"page",
							"total"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ListingCreateRequest": {
				"description": "Interface for creating a new listing. \ntsoa will use this to validate the POST body.",
				"properties": {
					"title": {
						"type": "string"
					},
					"address": {
						"type": "string",
						"minLength": 10
					},
					"price": {
						"type": "string"
					},
					"numOfBedrooms": {
						"type": "string"
					},
					"numOfBathrooms": {
						"type": "string"
					},
					"areaInSqFt": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"facilities": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"lat": {
						"type": "number",
						"format": "double",
						"description": "Latitude coordinate"
					},
					"lng": {
						"type": "number",
						"format": "double",
						"description": "Longitude coordinate"
					}
				},
				"required": [
					"title",
					"address",
					"price",
					"numOfBedrooms",
					"numOfBathrooms",
					"areaInSqFt",
					"images",
					"facilities",
					"lat",
					"lng"
				],
				"type": "object",
				"additionalProperties": true
			},
			"BookingCreateRequest": {
				"properties": {
					"listingId": {
						"type": "number",
						"format": "double"
					},
					"scheduledAt": {
						"type": "string",
						"format": "date-time"
					},
					"notes": {
						"type": "string"
					}
				},
				"required": [
					"listingId",
					"scheduledAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"BookingStatusUpdate": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"confirmed",
							"declined",
							"completed",
							"cancelled"
						]
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AuthResponse": {
				"properties": {
					"token": {
						"type": "string"
					},
					"user": {
						"properties": {
							"name": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"email",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"token",
					"user"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "real-state-api",
		"contact": {}
	},
	"paths": {
		"/upload/image": {
			"post": {
				"operationId": "UploadSingleImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"url": {
											"type": "string"
										}
									},
									"required": [
										"url"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Upload a single image for a listing.\nReturns the URL to be added to the listings.images array.",
				"tags": [
					"Media"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/reviews": {
			"post": {
				"operationId": "CreateReview",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Submit a review for a listing. \nPrevents duplicate reviews from the same user on the same listing.",
				"tags": [
					"Reviews"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewCreateRequest"
							}
						}
					}
				}
			}
		},
		"/reviews/listing/{listingId}": {
			"get": {
				"operationId": "GetListingReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get all reviews for a specific listing with User avatars and names.",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "listingId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reviews/listing/{listingId}/stats": {
			"get": {
				"operationId": "GetListingStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get overall rating stats for a listing.",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "listingId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/listings/search": {
			"get": {
				"operationId": "SearchListings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Search for listings within a specific radius (in meters)\nof a geographic point.",
				"tags": [
					"Real Estate"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "lat",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "lng",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "radius",
						"required": false,
						"schema": {
							"default": 5000,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/listings/paged": {
			"get": {
				"operationId": "GetPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_any_"
								}
							}
						}
					}
				},
				"description": "GET /listings/paged\nFetch listings with limit and offset pagination.",
				"tags": [
					"Listings"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/listings/featured": {
			"get": {
				"operationId": "GetFeatured",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "GET /listings/featured\nReturns top-tier or recently boosted listings.",
				"tags": [
					"Listings"
				],
				"security": [],
				"parameters": []
			}
		},
		"/listings/recommended": {
			"get": {
				"operationId": "GetRecommended",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "GET /listings/recommended\nUses spatial logic to find listings near a coordinate.",
				"tags": [
					"Listings"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "lat",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "lng",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/listings/my-listings": {
			"get": {
				"operationId": "GetMyListings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "GET /listings/my-listings\nReturns listings created by the authenticated user.",
				"tags": [
					"Listings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/listings/nearby": {
			"get": {
				"operationId": "GetNearbyListings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "GET /listings/nearby\nReturns listings near a given coordinate.",
				"tags": [
					"Listings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "lat",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "lng",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/listings/{id}": {
			"get": {
				"operationId": "GetListingDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Listing Detail",
				"tags": [
					"Listings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateListing",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "PUT /listings/{id}\nUpdate an existing listing.",
				"tags": [
					"Listings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteListing",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "DELETE /listings/{id}",
				"tags": [
					"Listings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/listings": {
			"post": {
				"operationId": "CreateListing",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"description": "Create a new listing.\nThe authenticated user's ID is automatically set as the agentId.",
				"tags": [
					"Listings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ListingCreateRequest"
							}
						}
					}
				}
			}
		},
		"/bookings": {
			"post": {
				"operationId": "CreateBooking",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Create a booking request.\nAutomatically assigns the agent from the listing.",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookingCreateRequest"
							}
						}
					}
				}
			}
		},
		"/bookings/my-requests": {
			"get": {
				"operationId": "GetMyRequests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get all bookings for the authenticated user (as a Buyer).",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/bookings/agent-schedule": {
			"get": {
				"operationId": "GetAgentSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get all bookings assigned to the authenticated user (as an Agent).",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/bookings/{id}/status": {
			"put": {
				"operationId": "UpdateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Update booking status (Confirm/Decline/Cancel).",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookingStatusUpdate"
							}
						}
					}
				}
			}
		},
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"avatar": {
												"type": "string"
											},
											"email": {
												"type": "string"
											},
											"name": {
												"type": "string"
											},
											"id": {
												"type": "string"
											}
										},
										"required": [
											"avatar",
											"email",
											"name",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/users/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"avatar": {
												"type": "string"
											},
											"email": {
												"type": "string"
											},
											"name": {
												"type": "string"
											},
											"id": {
												"type": "string"
											}
										},
										"required": [
											"avatar",
											"email",
											"name",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"phone": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"phone",
									"password",
									"email",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}